# Vectorize  {Last}IndexOfAny{Except} for ASCII needles