--- a/src/libraries/System.Memory/tests/Span/IndexOfAnyExcept.T.cs
+++ b/src/libraries/System.Memory/tests/Span/IndexOfAnyExcept.T.cs
@@ -55,6 +55,7 @@ namespace System.SpanTests
         }
 
         [Fact]
+        [OuterLoop("Takes about a second to execute")]
         public static void TestIndexOfAnyExcept_RandomInputs_Char()
         {
             IndexOfAnyCharTestHelper.TestRandomInputs(
@@ -76,6 +77,7 @@ namespace System.SpanTests
         }
 
         [Fact]
+        [OuterLoop("Takes about a second to execute")]
         public static void TestLastIndexOfAnyExcept_RandomInputs_Char()
         {
             IndexOfAnyCharTestHelper.TestRandomInputs(