… src/libraries/System.Memory/tests/Span/IndexOfAny.char.cs
… src/libraries/System.Memory/tests/Span/LastIndexOfAny.T.cs