Contributes to #68328 (and contributes the workhorse implementation for an eventual dedicated API)
cc: @gfoidl

This PR adds a vectorized path for `IndexOfAny`-like methods if the input is at least 8 characters long and the needle is only ASCII.
It is used if `Ssse3` or `AdvSimd.Arm64` are supported.

I also unified all `(Last)IndexOfAny(Except)` methods to use the same approach of
```c#
if (HaystackIsShort) return SimpleSearch();
if (NeedleIsAscii) return Vectorized();
return ProbabilisticMap();
```
leading to perf improvements for both short and long inputs.

For inputs with long runs of no matches, the vectorized path has 10 to 140+ times the throughput on realistic inputs.
As the `-Except` methods were previously `O(n * m)`, you can of course see arbitrarily large improvements. 

<details>
<summary><strong>X86 numbers</strong></summary>

These numbers were collected on a Windows Azure VM running on an Intel Xeon 8370C processor.

|               Method | Toolchain | Length |       Needle |         Mean |     Error | Ratio |
|--------------------- |---------- |------- |------------- |-------------:|----------:|------:|
|           IndexOfAny |      main |      1 |       ABCDEF |     9.787 ns | 0.0087 ns |  1.00 |
|           IndexOfAny |        pr |      1 |       ABCDEF |     5.852 ns | 0.0053 ns |  0.60 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |      1 |       ABCDEF |     5.313 ns | 0.0077 ns |  1.00 |
|     IndexOfAnyExcept |        pr |      1 |       ABCDEF |     6.676 ns | 0.0104 ns |  1.26 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |      1 |       ABCDEF |    11.376 ns | 0.0065 ns |  1.00 |
|       LastIndexOfAny |        pr |      1 |       ABCDEF |     5.783 ns | 0.0037 ns |  0.51 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |      1 |       ABCDEF |     5.946 ns | 0.0123 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |      1 |       ABCDEF |     6.030 ns | 0.0324 ns |  1.01 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |      1 | AlphaNumeric |    63.653 ns | 0.0265 ns |  1.00 |
|           IndexOfAny |        pr |      1 | AlphaNumeric |     5.855 ns | 0.0066 ns |  0.09 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |      1 | AlphaNumeric |     6.371 ns | 0.0128 ns |  1.00 |
|     IndexOfAnyExcept |        pr |      1 | AlphaNumeric |     7.619 ns | 0.0104 ns |  1.20 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |      1 | AlphaNumeric |    74.164 ns | 0.0712 ns |  1.00 |
|       LastIndexOfAny |        pr |      1 | AlphaNumeric |     5.558 ns | 0.0011 ns |  0.07 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |      1 | AlphaNumeric |     8.766 ns | 0.2798 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |      1 | AlphaNumeric |     6.859 ns | 0.0018 ns |  0.88 |
|                      |           |        |              |              |           |       |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |      8 |       ABCDEF |     14.12 ns |  0.016 ns |  1.00 |
|           IndexOfAny |        pr |      8 |       ABCDEF |     12.25 ns |  0.006 ns |  0.87 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |      8 |       ABCDEF |     24.38 ns |  0.019 ns |  1.00 |
|     IndexOfAnyExcept |        pr |      8 |       ABCDEF |     14.86 ns |  0.005 ns |  0.61 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |      8 |       ABCDEF |     16.51 ns |  0.008 ns |  1.00 |
|       LastIndexOfAny |        pr |      8 |       ABCDEF |     12.23 ns |  0.006 ns |  0.74 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |      8 |       ABCDEF |     23.42 ns |  0.022 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |      8 |       ABCDEF |     14.93 ns |  0.013 ns |  0.64 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |      8 | AlphaNumeric |     67.53 ns |  0.075 ns |  1.00 |
|           IndexOfAny |        pr |      8 | AlphaNumeric |     34.92 ns |  0.045 ns |  0.52 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |      8 | AlphaNumeric |     27.99 ns |  0.025 ns |  1.00 |
|     IndexOfAnyExcept |        pr |      8 | AlphaNumeric |     31.09 ns |  0.036 ns |  1.11 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |      8 | AlphaNumeric |     79.39 ns |  0.065 ns |  1.00 |
|       LastIndexOfAny |        pr |      8 | AlphaNumeric |     34.32 ns |  0.028 ns |  0.43 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |      8 | AlphaNumeric |     27.67 ns |  0.097 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |      8 | AlphaNumeric |     25.80 ns |  0.014 ns |  0.93 |
|                      |           |        |              |              |           |       |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |     16 |       ABCDEF |     20.15 ns |  0.060 ns |  1.00 |
|           IndexOfAny |        pr |     16 |       ABCDEF |     12.25 ns |  0.007 ns |  0.61 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |     16 |       ABCDEF |     43.59 ns |  0.043 ns |  1.00 |
|     IndexOfAnyExcept |        pr |     16 |       ABCDEF |     14.96 ns |  0.011 ns |  0.34 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |     16 |       ABCDEF |     21.76 ns |  0.012 ns |  1.00 |
|       LastIndexOfAny |        pr |     16 |       ABCDEF |     12.23 ns |  0.006 ns |  0.56 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |     16 |       ABCDEF |     41.38 ns |  0.033 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |     16 |       ABCDEF |     14.92 ns |  0.010 ns |  0.36 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |     16 | AlphaNumeric |     73.53 ns |  0.060 ns |  1.00 |
|           IndexOfAny |        pr |     16 | AlphaNumeric |     67.08 ns |  0.084 ns |  0.91 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |     16 | AlphaNumeric |     49.79 ns |  0.032 ns |  1.00 |
|     IndexOfAnyExcept |        pr |     16 | AlphaNumeric |     54.95 ns |  0.051 ns |  1.10 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |     16 | AlphaNumeric |     84.76 ns |  0.061 ns |  1.00 |
|       LastIndexOfAny |        pr |     16 | AlphaNumeric |     66.56 ns |  0.060 ns |  0.79 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |     16 | AlphaNumeric |     50.64 ns |  0.162 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |     16 | AlphaNumeric |     45.80 ns |  0.062 ns |  0.91 |
|                      |           |        |              |              |           |       |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |     32 |       ABCDEF |     29.04 ns |  0.034 ns |  1.00 |
|           IndexOfAny |        pr |     32 |       ABCDEF |     13.37 ns |  0.009 ns |  0.46 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |     32 |       ABCDEF |     83.02 ns |  0.093 ns |  1.00 |
|     IndexOfAnyExcept |        pr |     32 |       ABCDEF |     15.98 ns |  0.007 ns |  0.19 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |     32 |       ABCDEF |     31.55 ns |  0.022 ns |  1.00 |
|       LastIndexOfAny |        pr |     32 |       ABCDEF |     13.92 ns |  0.050 ns |  0.44 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |     32 |       ABCDEF |     82.09 ns |  0.090 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |     32 |       ABCDEF |     15.77 ns |  0.007 ns |  0.19 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |     32 | AlphaNumeric |     83.50 ns |  0.061 ns |  1.00 |
|           IndexOfAny |        pr |     32 | AlphaNumeric |     65.77 ns |  0.252 ns |  0.79 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |     32 | AlphaNumeric |    106.94 ns |  0.054 ns |  1.00 |
|     IndexOfAnyExcept |        pr |     32 | AlphaNumeric |     66.03 ns |  0.099 ns |  0.62 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |     32 | AlphaNumeric |     94.24 ns |  0.052 ns |  1.00 |
|       LastIndexOfAny |        pr |     32 | AlphaNumeric |     64.81 ns |  0.073 ns |  0.69 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |     32 | AlphaNumeric |    107.73 ns |  0.249 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |     32 | AlphaNumeric |     65.58 ns |  0.034 ns |  0.61 |
|                      |           |        |              |              |           |       |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |     64 |       ABCDEF |     51.95 ns |  0.321 ns |  1.00 |
|           IndexOfAny |        pr |     64 |       ABCDEF |     15.31 ns |  0.009 ns |  0.30 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |     64 |       ABCDEF |    165.25 ns |  0.166 ns |  1.00 |
|     IndexOfAnyExcept |        pr |     64 |       ABCDEF |     17.54 ns |  0.026 ns |  0.11 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |     64 |       ABCDEF |     52.16 ns |  0.039 ns |  1.00 |
|       LastIndexOfAny |        pr |     64 |       ABCDEF |     15.83 ns |  0.020 ns |  0.30 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |     64 |       ABCDEF |    164.63 ns |  0.167 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |     64 |       ABCDEF |     17.01 ns |  0.008 ns |  0.10 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |     64 | AlphaNumeric |    105.16 ns |  0.079 ns |  1.00 |
|           IndexOfAny |        pr |     64 | AlphaNumeric |     67.26 ns |  0.093 ns |  0.64 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |     64 | AlphaNumeric |    208.43 ns |  0.070 ns |  1.00 |
|     IndexOfAnyExcept |        pr |     64 | AlphaNumeric |     68.18 ns |  0.107 ns |  0.33 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |     64 | AlphaNumeric |    115.14 ns |  0.057 ns |  1.00 |
|       LastIndexOfAny |        pr |     64 | AlphaNumeric |     66.35 ns |  0.040 ns |  0.58 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |     64 | AlphaNumeric |    212.70 ns |  0.583 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |     64 | AlphaNumeric |     67.31 ns |  0.059 ns |  0.32 |
|                      |           |        |              |              |           |       |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |    128 |       ABCDEF |     90.31 ns |  0.083 ns |  1.00 |
|           IndexOfAny |        pr |    128 |       ABCDEF |     18.74 ns |  0.008 ns |  0.21 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |    128 |       ABCDEF |    336.07 ns |  0.234 ns |  1.00 |
|     IndexOfAnyExcept |        pr |    128 |       ABCDEF |     20.78 ns |  0.015 ns |  0.06 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |    128 |       ABCDEF |    100.96 ns |  0.050 ns |  1.00 |
|       LastIndexOfAny |        pr |    128 |       ABCDEF |     18.80 ns |  0.024 ns |  0.19 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |    128 |       ABCDEF |    337.61 ns |  0.233 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |    128 |       ABCDEF |     21.00 ns |  0.040 ns |  0.06 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |    128 | AlphaNumeric |    148.35 ns |  0.231 ns |  1.00 |
|           IndexOfAny |        pr |    128 | AlphaNumeric |     70.20 ns |  0.121 ns |  0.47 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |    128 | AlphaNumeric |    425.60 ns |  0.178 ns |  1.00 |
|     IndexOfAnyExcept |        pr |    128 | AlphaNumeric |     71.25 ns |  0.041 ns |  0.17 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |    128 | AlphaNumeric |    165.64 ns |  0.085 ns |  1.00 |
|       LastIndexOfAny |        pr |    128 | AlphaNumeric |     70.08 ns |  0.118 ns |  0.42 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |    128 | AlphaNumeric |    438.12 ns |  1.407 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |    128 | AlphaNumeric |     72.16 ns |  0.126 ns |  0.16 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |    256 |       ABCDEF |    182.60 ns |  0.382 ns |  1.00 |
|           IndexOfAny |        pr |    256 |       ABCDEF |     29.07 ns |  0.026 ns |  0.16 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |    256 |       ABCDEF |    703.61 ns |  0.468 ns |  1.00 |
|     IndexOfAnyExcept |        pr |    256 |       ABCDEF |     32.34 ns |  0.019 ns |  0.05 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |    256 |       ABCDEF |    196.47 ns |  1.878 ns |  1.00 |
|       LastIndexOfAny |        pr |    256 |       ABCDEF |     29.52 ns |  0.063 ns |  0.15 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |    256 |       ABCDEF |    704.98 ns |  0.586 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |    256 |       ABCDEF |     31.83 ns |  0.019 ns |  0.05 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |    256 | AlphaNumeric |    236.97 ns |  0.128 ns |  1.00 |
|           IndexOfAny |        pr |    256 | AlphaNumeric |     81.93 ns |  0.287 ns |  0.35 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |    256 | AlphaNumeric |    878.10 ns |  0.445 ns |  1.00 |
|     IndexOfAnyExcept |        pr |    256 | AlphaNumeric |     82.96 ns |  0.094 ns |  0.09 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |    256 | AlphaNumeric |    247.84 ns |  0.110 ns |  1.00 |
|       LastIndexOfAny |        pr |    256 | AlphaNumeric |     80.38 ns |  0.055 ns |  0.32 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |    256 | AlphaNumeric |    871.79 ns |  0.353 ns |  1.00 |
| LastIndexOfAnyExcept |        pr |    256 | AlphaNumeric |     83.21 ns |  0.096 ns |  0.10 |
|                      |           |        |              |              |           |       |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |  10000 |       ABCDEF |   6,409.6 ns |   4.42 ns |  1.00 |
|           IndexOfAny |        pr |  10000 |       ABCDEF |     651.2 ns |   0.36 ns |  0.10 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |  10000 |       ABCDEF |  71,960.2 ns |  31.30 ns |  1.00 |
|     IndexOfAnyExcept |        pr |  10000 |       ABCDEF |     741.1 ns |   0.36 ns |  0.01 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |  10000 |       ABCDEF |   6,396.7 ns |   2.27 ns |  1.00 |
|       LastIndexOfAny |        pr |  10000 |       ABCDEF |     643.2 ns |   0.36 ns |  0.10 |
|                      |           |        |              |              |           |       |
| LastIndexOfAnyExcept |      main |  10000 |       ABCDEF |  81,927.2 ns |  23.11 ns | 1.000 |
| LastIndexOfAnyExcept |        pr |  10000 |       ABCDEF |     728.1 ns |   0.26 ns | 0.009 |
|                      |           |        |              |              |           |       |
|           IndexOfAny |      main |  10000 | AlphaNumeric |   6,450.1 ns |   2.54 ns |  1.00 |
|           IndexOfAny |        pr |  10000 | AlphaNumeric |     704.2 ns |   0.41 ns |  0.11 |
|                      |           |        |              |              |           |       |
|     IndexOfAnyExcept |      main |  10000 | AlphaNumeric | 114,854.9 ns |  91.93 ns | 1.000 |
|     IndexOfAnyExcept |        pr |  10000 | AlphaNumeric |     791.0 ns |   0.36 ns | 0.007 |
|                      |           |        |              |              |           |       |
|       LastIndexOfAny |      main |  10000 | AlphaNumeric |   6,613.7 ns |  21.32 ns |  1.00 |
|       LastIndexOfAny |        pr |  10000 | AlphaNumeric |     690.0 ns |   0.08 ns |  0.10 |
…