@stephentoub 2022-10-07
Excellent, thanks for working on this.

> The (Last)IndexOfAnyExcept methods that were previously using an O(n * m) loop regress with this change if the needle is long, while the haystack is small

Have we investigated doing something different for short haystacks? e.g. something super simple like:
```C#
if (haystack.Length < Vector128<short>.Count)
{
    for (int i = 0; i < haystack.Length; i++)
        if (needle.Contains(haystack[i]))
            return i;
    return -1;
}
...
```
?