@MihaZupan 2022-10-07
That is what we do for the `-Except` overloads with this PR.
```c#
haystack.Length >= 8 ? Vectorized() : SimpleLoop()
```
whereas `(Last)IndexOfAny` does
```c#
haystack.Length >= 8 ? Vectorized() : ProbabilisticMap()
```
In places where we were already using the `ProbabilisticMap`, the init overhead seems to be very similar.

It would likely be beneficial to tweak the exact cutoff and add the `SimpleLoop` variant where we currently only use the `ProbabilisticMap`.
E.g. (made up constants)
```c#
int IndexOfAny()
{
    if (haystack.Length < Max(8, needle.Length / 2))
        return SimpleLoop();

    if (IsAscii(needle))
        return Vectorized();

    return ProbabilisticMap();
}

int IndexOfAnyExcept()
{
    if (haystack.Length >= Max(8, needle.Length / 2) && IsAscii(needle))
        return Vectorized();

    return SimpleLoop();
}
```

I can look into what sort of numbers we'd see with something like that, though I would hope that in general, a dedicated API to hide the init cost completely would be the preferred approach.