@MihaZupan 2022-10-08
Rerunning the benchmarks with https://github.com/MihaZupan/runtime/commit/e06c007df6552892f642c2a8d1bdafb60be3892b now.