@MihaZupan 2022-10-09
I added a fast path for short haystacks to the probabilistic code path, leading to nice improvements there (I updated the numbers in the top post). E.g.
|               Method | Toolchain | Length |       Needle |         Mean |     Error | Ratio |
|--------------------- |---------- |------- |------------- |-------------:|----------:|------:|
|       LastIndexOfAny |      main |      1 | AlphaNumeric |    74.164 ns | 0.0712 ns |  1.00 |
|       LastIndexOfAny |        pr |      1 | AlphaNumeric |     5.558 ns | 0.0011 ns |  0.07 |

The only real regression is the `(Last)IndexOfAnyExcept` for long needles and short haystacks as mentioned before. I'll keep playing around with cutoffs to try and improve that case too.