@MihaZupan 2022-10-09
The cutoff for the simple loop is now:
```c#
searchSpaceLength < Vector128<short>.Count || (searchSpaceLength < 20 && searchSpaceLength < (valuesLength >> 1))
```

Updated the benchmarks above, this is now a win pretty much across the board.