@danmoseley 2022-10-10
Curious to see what this does to regex redux 

https://benchmarksgame-team.pages.debian.net/benchmarksgame/program/regexredux-csharpcore-5.html