@stephentoub 2022-10-10
> Curious to see what this does to regex redux

It won't do anything.  We don't use this API from RegexCompiler / source generator, nor does regex redux have any sets that would trigger it.  We _will_ update RegexCompiler / source generator to use whatever new API we create that plugs into the same implementation this is adding but that let's us precompute the vector/bit set rather than having to do it on each call.