@MihaZupan 2022-11-02
> Meaning with this PR, for long or short inputs, vectorized or not, ASCII or not, everything you've tried is as good or better than it was before? If so, yay!

Yes.

The main scenario I can think of that would regress is if the needle contains non-ASCII, but starts with ASCII (e.g. `alphabet + 'ü'`) as you now have to pay for an extra scan to build the ASCII bitmap that you don't actually need.